+++ example/android/app/src/main/kotlin/com/cavdarfurkan/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.cavdarfurkan.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()